Move config to scanner package
