+++ core-lib/src/main/java/org/idpass/smartscanner/lib/scanner/config/Config.kt
@@ -15,26 +15,29 @@
  *
  *
  */
-package org.idpass.smartscanner.lib.config
+package org.idpass.smartscanner.lib.scanner.config
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import org.idpass.smartscanner.lib.platform.extension.empty
 
 @Parcelize
-data class BarcodeOptions(
-    val barcodeFormats: List<String>? = null,
-    val barcodeScannerSize: String? = null,
-    val idPassLiteSupport: Boolean? = null
+data class Config(
+    val background: String? = null,
+    val font: String? = null,
+    val imageResultType: String? = null,
+    val label: String? = null,
+    val isManualCapture: Boolean? = null,
+    val branding: Boolean? = null,
 ) : Parcelable {
     companion object {
-        val default = BarcodeOptions(
-            barcodeFormats = BarcodeFormat.default,
-            idPassLiteSupport = false
-        )
-        val defaultIdPassLite = BarcodeOptions(
-            barcodeScannerSize = ScannerSize.LARGE.value,
-            barcodeFormats = arrayListOf(BarcodeFormat.QR_CODE.label),
-            idPassLiteSupport = true
+        val default = Config(
+            background = String.empty(),
+            font = String.empty(),
+            imageResultType = ImageResultType.PATH.value,
+            label = String.empty(),
+            isManualCapture = false,
+            branding = true
         )
     }
 }
